brightness: follow custom volume control themes
follow custom volume control themes. Brightness slider a near match
for volume control slider by using same .mate-panel-applet-slider
selector as in mate-media.
One difference is that this slider is drawn in a single frame,
volume control slider is in a frame within a frame.
Therefore this one is drawn slightly narrower.