brightness: follow custom volume control themes
brightness: follow custom volume control themes. Brightness slider a near match for volume control slider by using same .mate-media-applet-slider selector as in mate-media. One difference is that this slider is drawn in a single frame, volume control slider is in a frame within a frame. Therefore this one is drawn slightly narrower
Coding note: most code near the patch uses tabs but a few lines use spaces. Tabs used for consistancy with code below patch and most of code in the file