engine: Use primary instead of composite when n<=1
With the release of Linux 4.17, the kernel will support the setting
of battery charge thresholds on Lenovo ThinkPad laptops.

This means, that some of the batteries can be in a state of "Not
Charging", which is new to the ACPI battery driver.

When in such a state, GPM reports a wrong icon in the tray
(battery-missing-symbolic) because the state on the fake composite
device is not updated.

This patch returns the actual device if the number of batteries is
only 1 instead of the composite device. This fixes the icon in the
tray.

If the number of batteries is greater than 1, the fake composite
device is returned and everyhting resumes business as usual.

Signed-off-by: Ognjen Galic <smclt30p@gmail.com>