brightness applet: Prevent sliding from bogging down the UI
Do not perform a synchronous DBus call to update the brightness in real
time when the slider move, because at least on some hardware the call
is slow, which leads to the UI bogging down and looking frozen.

To fix this, cap the update rate in response to slider motion.  This
makes the brightness apply slightly slower, but prevent most bogging
down (if the rate is well chosen, at least).