Refactor: kbd_backlight: Backlight toggling
This improves the fix for backlight toggling from "Keyboard backlight
handling improvements".
Now saved_brightness is initilized to 0 and only gets set to current
brightness when "toggling off" - which is the only place where setting
saved_brightness to current brightness makes any sense, logically. In
all other cases it's not needed.

I adapted the code that would crank up the brightness to 100 if we
"toggle on" to 0 although the "arbitrary" special handling introduced
by that is what I personally would consider broken.