gpm-kbd-backlight: Persistently save user-set brightness
Try and save the user-set brightness to be able to restore it on next
restart.

This is not perfect because we have complex interaction between
settings which makes it not trivial, if at all possible, to save the
value yet respect the dim settings.

Here we try to save the value in a way that counteract battery dimming
so that restoring the value yields expected results (e.g. a value saved
on battery restores identical on battery again), but it cannot work
when the value to save is larger than the dimmed maximum as at the
moment the maximum value is 100%.

It is however the best I could come up with working with the settings
we currently have, and it's probably good enough at least as a starting
point.

It also switches from `master_percentage` to `brightness_percent` in
the toggle code because the former is only initialized to the actual
current brightness value once brightness changed at least once, while
the former is properly initialized.  It is otherwise quivalent for this
feature, and less confusing as to when the value gets updated.